requirements.txt
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0
requests==2.31.0
ccxt==3.1.43
matplotlib==3.7.2
ta==0.10.2
python-dateutil==2.8.2
pytz==2023.3
tabulate==0.9.0
websocket-client==1.6.1
python-binance==1.0.19
